# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Portfolio API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ReadinessStatusStatus = typeof ReadinessStatusStatus[keyof typeof ReadinessStatusStatus];


export const ReadinessStatusStatus = {
  ready: 'ready',
  not_ready: 'not_ready',
} as const;

export interface ReadinessStatus {
  status: ReadinessStatusStatus;
}

export interface ContactInput {
  /**
     * @minLength 2
     * @maxLength 80
     */
  name: string;
  /** @maxLength 160 */
  email: string;
  /**
     * @minLength 3
     * @maxLength 120
     */
  subject: string;
  /**
     * @minLength 20
     * @maxLength 3000
     */
  message: string;
}

export interface ContactResult {
  ok: boolean;
}

export interface PortfolioVisitInput {
  visitorId: string;
}

export interface PortfolioStats {
  /** @minimum 0 */
  visitors: number;
  /** @minimum 0 */
  inquiries: number;
  /** @minimum 0 */
  activeBuilds: number;
}

export interface ChatInput {
  /**
     * @minLength 2
     * @maxLength 800
     */
  message: string;
}

export interface ChatResponse {
  /**
     * @minLength 1
     * @maxLength 1600
     */
  answer: string;
  /**
     * @minItems 0
     * @maxItems 4
     * @items.minLength 1
     * @items.maxLength 120
     */
  suggestedQuestions: string[];
}

export interface ApiError {
  error: string;
}