# packages/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Portfolio API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  /**
     * @minLength 2
     * @maxLength 80
     */
  name: string;
  /** @maxLength 160 */
  email: string;
  /**
     * @minLength 3
     * @maxLength 120
     */
  subject: string;
  /**
     * @minLength 20
     * @maxLength 3000
     */
  message: string;
}